refactor: extract logic detection and source parsing into helpers
LogicDetector owns the branching-node list and the recursive walker as
static methods. SourceParser owns file parsing with the trait-method
and use-map caches. Rule file drops from ~440 to ~220 lines.
