+++ src/Core/DevOps/StaticAnalyze/PHPStan/Rules/CodeCoverageIgnoreEvaluationRule.php
@@ -3,19 +3,14 @@
 namespace Shopware\Core\DevOps\StaticAnalyze\PHPStan\Rules;
 
 use PhpParser\Node;
-use PhpParser\Node\Expr;
-use PhpParser\Node\Stmt;
 use PhpParser\Node\Stmt\Class_;
-use PhpParser\Node\Stmt\ClassMethod;
-use PhpParser\Node\Stmt\Trait_;
-use PhpParser\NodeFinder;
-use PhpParser\Parser;
-use PhpParser\ParserFactory;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Rules\IdentifierRuleError;
 use PHPStan\Rules\Rule;
 use Shopware\Core\DevOps\StaticAnalyze\PHPStan\Rules\CodeCoverageIgnore\Errors;
+use Shopware\Core\DevOps\StaticAnalyze\PHPStan\Rules\CodeCoverageIgnore\LogicDetector;
+use Shopware\Core\DevOps\StaticAnalyze\PHPStan\Rules\CodeCoverageIgnore\SourceParser;
 use Shopware\Core\Framework\Log\Package;
 
 /**
@@ -26,52 +21,12 @@ use Shopware\Core\Framework\Log\Package;
 #[Package('framework')]
 class CodeCoverageIgnoreEvaluationRule implements Rule
 {
-    /**
-     * What "logic" means for this rule: anything that introduces a branch or an
-     * error path. Calls, instantiation, arithmetic, and coalesce are intentionally
-     * absent — they're not branching by themselves, and the called code has its
-     * own coverage story.
-     *
-     * @var list<class-string<Node>>
-     */
-    private const LOGIC_NODE_TYPES = [
-        Stmt\If_::class,
-        Stmt\ElseIf_::class,
-        Stmt\Else_::class,
-        Stmt\Switch_::class,
-        Expr\Match_::class,
-        Stmt\While_::class,
-        Stmt\Do_::class,
-        Stmt\For_::class,
-        Stmt\Foreach_::class,
-        Stmt\TryCatch::class,
-        Stmt\Catch_::class,
-        Stmt\Throw_::class,
-        Expr\Throw_::class,
-        Expr\Ternary::class,
-    ];
-
-    private readonly Parser $parser;
-
-    private readonly NodeFinder $finder;
-
-    /**
-     * @var array<string, list<ClassMethod>>
-     */
-    private array $traitMethodCache = [];
-
-    /**
-     * file path => alias => FQCN
-     *
-     * @var array<string, array<string, string>>
-     */
-    private array $useMapCache = [];
+    private readonly SourceParser $sources;
 
     public function __construct(
         private readonly ReflectionProvider $reflectionProvider,
     ) {
-        $this->parser = (new ParserFactory())->createForHostVersion();
-        $this->finder = new NodeFinder();
+        $this->sources = new SourceParser($reflectionProvider);
     }
 
     public function getNodeType(): string
@@ -129,7 +84,7 @@ class CodeCoverageIgnoreEvaluationRule implements Rule
         foreach ($node->getMethods() as $method) {
             $methodName = (string) $method->name;
 
-            if ($classHasIgnore && !$classExempted && $this->methodContainsLogic($method)) {
+            if ($classHasIgnore && !$classExempted && LogicDetector::methodContainsLogic($method)) {
                 $errors[] = Errors::classLevel($className, $methodName, $method->getStartLine());
 
                 continue;
@@ -143,7 +98,7 @@ class CodeCoverageIgnoreEvaluationRule implements Rule
                 continue;
             }
 
-            if ($this->methodContainsLogic($method)) {
+            if (LogicDetector::methodContainsLogic($method)) {
                 $errors[] = Errors::methodLevel($className, $methodName, $method->getStartLine());
             }
         }
@@ -166,17 +121,22 @@ class CodeCoverageIgnoreEvaluationRule implements Rule
 
         $errors = [];
 
-        foreach ($this->traitMethods($node) as [$traitName, $method]) {
-            if (!$this->methodContainsLogic($method)) {
-                continue;
-            }
+        foreach ($node->getTraitUses() as $use) {
+            foreach ($use->traits as $traitName) {
+                $name = $traitName->toString();
+                foreach ($this->sources->traitMethods($name) as $method) {
+                    if (!LogicDetector::methodContainsLogic($method)) {
+                        continue;
+                    }
 
-            $errors[] = Errors::traitMethod(
-                $className,
-                $traitName,
-                (string) $method->name,
-                $node->getStartLine(),
-            );
+                    $errors[] = Errors::traitMethod(
+                        $className,
+                        $name,
+                        (string) $method->name,
+                        $node->getStartLine(),
+                    );
+                }
+            }
         }
 
         return $errors;
@@ -223,7 +183,7 @@ class CodeCoverageIgnoreEvaluationRule implements Rule
             // use statements. Qualified refs (with `\` or relative path) are
             // taken as-is, matching common phpdoc conventions in this codebase.
             if (!str_starts_with($rawClass, '\\') && !str_contains($candidate, '\\')) {
-                $useMap ??= $this->getUseMap($scope->getFile());
+                $useMap ??= $this->sources->useMap($scope->getFile());
                 $resolved = $useMap[$candidate] ?? $candidate;
             }
 
@@ -237,71 +197,6 @@ class CodeCoverageIgnoreEvaluationRule implements Rule
         }
 
         return false;
-    }
-
-    /**
-     * @return array<string, string>
-     */
-    private function getUseMap(string $file): array
-    {
-        if (\array_key_exists($file, $this->useMapCache)) {
-            return $this->useMapCache[$file];
-        }
-
-        $this->useMapCache[$file] = [];
-
-        if (!is_file($file)) {
-            return [];
-        }
-
-        $source = @file_get_contents($file);
-        if ($source === false) {
-            return [];
-        }
-
-        try {
-            $stmts = $this->parser->parse($source);
-        } catch (\Throwable) {
-            return [];
-        }
-
-        if ($stmts === null) {
-            return [];
-        }
-
-        $map = [];
-        foreach ($stmts as $stmt) {
-            if ($stmt instanceof Stmt\Namespace_) {
-                foreach ($stmt->stmts as $inner) {
-                    $this->collectUses($inner, $map);
-                }
-            } else {
-                $this->collectUses($stmt, $map);
-            }
-        }
-
-        return $this->useMapCache[$file] = $map;
-    }
-
-    /**
-     * @param array<string, string> $map
-     */
-    private function collectUses(Node $stmt, array &$map): void
-    {
-        if ($stmt instanceof Stmt\Use_) {
-            foreach ($stmt->uses as $use) {
-                $map[$use->getAlias()->name] = $use->name->toString();
-            }
-
-            return;
-        }
-
-        if ($stmt instanceof Stmt\GroupUse) {
-            $prefix = $stmt->prefix->toString();
-            foreach ($stmt->uses as $use) {
-                $map[$use->getAlias()->name] = $prefix . '\\' . $use->name->toString();
-            }
-        }
     }
 
     private function className(Class_ $node): string
@@ -321,118 +216,5 @@ class CodeCoverageIgnoreEvaluationRule implements Rule
         }
 
         return (bool) preg_match('/@codeCoverageIgnore(?![A-Za-z])/', $doc->getText());
-    }
-
-    private function methodContainsLogic(ClassMethod $method): bool
-    {
-        if ($method->stmts === null) {
-            return false;
-        }
-
-        foreach ($method->stmts as $stmt) {
-            if ($this->nodeContainsLogic($stmt)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private function nodeContainsLogic(Node $node): bool
-    {
-        foreach (self::LOGIC_NODE_TYPES as $type) {
-            if ($node instanceof $type) {
-                return true;
-            }
-        }
-
-        foreach ($node->getSubNodeNames() as $name) {
-            $value = $node->{$name};
-            if ($value instanceof Node) {
-                if ($this->nodeContainsLogic($value)) {
-                    return true;
-                }
-            } elseif (\is_array($value)) {
-                foreach ($value as $item) {
-                    if ($item instanceof Node && $this->nodeContainsLogic($item)) {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * @return iterable<array{0: string, 1: ClassMethod}>
-     */
-    private function traitMethods(Class_ $class): iterable
-    {
-        foreach ($class->getTraitUses() as $use) {
-            foreach ($use->traits as $traitName) {
-                $name = $traitName->toString();
-                foreach ($this->loadTraitMethods($name) as $method) {
-                    yield [$name, $method];
-                }
-            }
-        }
-    }
-
-    /**
-     * @return list<ClassMethod>
-     */
-    private function loadTraitMethods(string $traitName): array
-    {
-        if (\array_key_exists($traitName, $this->traitMethodCache)) {
-            return $this->traitMethodCache[$traitName];
-        }
-
-        $this->traitMethodCache[$traitName] = [];
-
-        if (!$this->reflectionProvider->hasClass($traitName)) {
-            return [];
-        }
-
-        $reflection = $this->reflectionProvider->getClass($traitName);
-        if (!$reflection->isTrait()) {
-            return [];
-        }
-
-        $file = $reflection->getFileName();
-        if ($file === null || !is_file($file)) {
-            return [];
-        }
-
-        $source = @file_get_contents($file);
-        if ($source === false) {
-            return [];
-        }
-
-        try {
-            $stmts = $this->parser->parse($source);
-        } catch (\Throwable) {
-            return [];
-        }
-
-        if ($stmts === null) {
-            return [];
-        }
-
-        $shortName = ($pos = strrpos($traitName, '\\')) === false ? $traitName : substr($traitName, $pos + 1);
-
-        /** @var list<Trait_> $traitNodes */
-        $traitNodes = $this->finder->findInstanceOf($stmts, Trait_::class);
-        foreach ($traitNodes as $traitNode) {
-            if ($traitNode->name === null || $traitNode->name->name !== $shortName) {
-                continue;
-            }
-
-            $this->traitMethodCache[$traitName] = $traitNode->getMethods();
-
-            return $this->traitMethodCache[$traitName];
-        }
-
-        return [];
     }
 }
